Decoding to Unicode prior to Slumber ``.post``.
Doing this in ``copy_files`` of lib/storageService.py fixes a bug where
Latin-1-encodable directory names are (seemingly) incorrectly
base-64-encoded, causing Slumber's ``.post`` call to raise a
UnicodeDecodeError, complaining of invalid continuation bytes. To
recreate the bug in commits prior to this one, attempt to transfer in a
directory whose name contains an accutely accented vowel. Interestingly,
the transfer works if you add a character to the directory name which is
only encodable in a Unicode-compatible codec, e.g., U+0294.